Trying to switch a bit

+++ src/responsibleai/rai_analyse/deployed_model_loader.py
@@ -4,12 +4,28 @@
 
 import json
 import logging
+import os
 import subprocess
+import tempfile
 import time
 
 import requests
 
 import pandas as pd
+
+from azure.identity import DefaultAzureCredential
+
+from azureml.core import Workspace
+
+from rai_component_utilities import (
+    print_dir_tree,
+    load_dataset,
+    fetch_model_id,
+    load_mlflow_model,
+    download_model_to_dir,
+)
+
+from model_wrapper import ModelWrapper
 
 _logger = logging.getLogger(__file__)
 logging.basicConfig(level=logging.INFO)
@@ -29,11 +45,36 @@ def check_for_ready_string(log_output: str):
     return result
 
 
-class DeployedModel:
-    def __init__(self, model_dir: str):
-        self._target_model_dir = model_dir
+class DeployedModelLoader:
+    def __init__(self, workspace: Workspace, model_id: str):
+        self._sub_id = workspace.subscription_id
+        self._resource_group = workspace.resource_group
+        self._workspace_name = workspace._workspace_name
+        self._model_id = model_id
+        self._unwrapped_model_dir = tempfile.mkdtemp()
 
     def __enter__(self):
+        self._server = None
+
+        return self
+
+    def load(self, path: str):
+        _logger.info(f"Ignoring supplied path: {path}")
+        _logger.info("Creating workspace object")
+        workspace = Workspace(self._sub_id, self._resource_group, self._workspace_name, auth=DefaultAzureCredential(exclude_shared_token_cache_credential=True))
+
+        _logger.info("Downloading mlflow model from AzureML")
+        download_model_to_dir(
+            workspace, self._model_id, self._unwrapped_model_dir
+        )
+        model_name = self._model_id.split(":")[0]
+
+        _logger.info("Trying to create wrapped model")
+        self._target_model_dir = ModelWrapper.wrap_mlflow_model(
+            os.path.join(self._unwrapped_model_dir, model_name)
+        )
+
+        _logger.info("Starting mlflow process")
         launch_args = [
             "mlflow",
             "models",
@@ -55,17 +96,23 @@ class DeployedModel:
             if self._server.returncode is not None:
                 # Server has crashed
                 raise RuntimeError("MLFlow server has crashed")
+        _logger.info("Server coming up.... pausing")
+        time.sleep(10)
 
         _logger.info("MLFlow model deployed")
         return self
 
     def __exit__(self, exception_type, exception_value, exception_traceback):
-        _logger.info("Sending SIGTERM to server process")
-        self._server.terminate()
-        time.sleep(5)
-        _logger.info("Sending SIGKILL to server process")
-        self._server.kill()
-        _logger.info("Process killed")
+        if self._server is not None:
+            _logger.info("Sending SIGTERM to server process")
+            self._server.terminate()
+            time.sleep(5)
+            _logger.info("Sending SIGKILL to server process")
+            self._server.kill()
+            _logger.info("Process killed")
+        else:
+            _logger.info("No server found")
+        
 
     def _call_model_and_extract(self, input_df: pd.DataFrame, target: str):
         payload = input_df.to_json(orient="split")
